Updated to 5.0

+++ php/settings/ajax/togglegroups.php
@@ -1,29 +1,42 @@
 <?php
 
-// Init owncloud
-require_once('../../lib/base.php');
-
-OC_JSON::checkAdminUser();
+OC_JSON::checkSubAdminUser();
 OCP\JSON::callCheck();
 
 $success = true;
-$error = "add user to";
-$action = "add";
-
 $username = $_POST["username"];
-$group = htmlentities($_POST["group"]);
+$group = $_POST["group"];
 
-if(!OC_Group::groupExists($group)){
+if($username == OC_User::getUser() && $group == "admin" &&  OC_User::isAdminUser($username)) {
+	$l = OC_L10N::get('core');
+	OC_JSON::error(array( 'data' => array( 'message' => $l->t('Admins can\'t remove themself from the admin group'))));
+	exit();
+}
+
+if(!OC_User::isAdminUser(OC_User::getUser())
+	&& (!OC_SubAdmin::isUserAccessible(OC_User::getUser(), $username)
+		|| !OC_SubAdmin::isGroupAccessible(OC_User::getUser(), $group))) {
+	$l = OC_L10N::get('core');
+	OC_JSON::error(array( 'data' => array( 'message' => $l->t('Authentication error') )));
+	exit();
+}
+
+if(!OC_Group::groupExists($group)) {
 	OC_Group::createGroup($group);
 }
 
+$l = OC_L10N::get('settings');
+
+$error = $l->t("Unable to add user to group %s", $group);
+$action = "add";
+
 // Toggle group
-if( OC_Group::inGroup( $username, $group )){
+if( OC_Group::inGroup( $username, $group )) {
 	$action = "remove";
-	$error = "remove user from";
+	$error = $l->t("Unable to remove user from group %s", $group);
 	$success = OC_Group::removeFromGroup( $username, $group );
 	$usersInGroup=OC_Group::usersInGroup($group);
-	if(count($usersInGroup)==0){
+	if(count($usersInGroup)==0) {
 		OC_Group::deleteGroup($group);
 	}
 }
@@ -32,11 +45,9 @@ else{
 }
 
 // Return Success story
-if( $success ){
+if( $success ) {
 	OC_JSON::success(array("data" => array( "username" => $username, "action" => $action, "groupname" => $group )));
 }
 else{
-	OC_JSON::error(array("data" => array( "message" => "Unable to $error group $group" )));
+	OC_JSON::error(array("data" => array( "message" => $error )));
 }
-
-?>
